chore(cachedb): move Cache interface to pkg/storage/types
